Adding Error handling

+++ db/db_user.py
@@ -1,4 +1,5 @@
 #ORM functionality
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
 
 from db.hash import Hash
@@ -21,12 +22,15 @@ def get_all_users(db: Session):
     return db.query(DbUser).all()
 
 def get_user(db: Session, id: int):
-    # Handle any exceptions here
-    return db.query(DbUser).filter(DbUser.id == id).first() # filters output based on id
+    user = db.query(DbUser).filter(DbUser.id == id).first() # filters output based on id
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
+    return user
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
-    # Handle any exceptions here
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -37,7 +41,8 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
-    # Handle any exceptions here
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
     db.delete(user)
     db.commit()
     return "deleted"
